// src/WebClient/mssql.client/src/app/table/models/TableColumn.ts
export interface TableColumn {
    tableName: string;
    columnName: string;
    key: string;
    identity: string;
    dataType: string;
    maxLength: string;
    allowNulls: string;
    default: string;
    description: string;
    descriptionname: string;
    extendedPropertyName: string;
    extendedPropertyValue: string;
}